Escape page titles
Any HTML tags in the titles were sent verbatim to the client. This
breaks a bunch of stuff.

It's also a potential XSS vulnerability, but since (a) I'm the only one
posting content -- via Github; and (b) I can do that via the page
content; or (c) because I own the templates, it's low-risk.